Move JointRig into joint.py
