+++ services/ucs/tests_ucs/fixtures/joint.py
@@ -21,11 +21,13 @@ from collections.abc import AsyncGenerator
 from dataclasses import dataclass
 
 import httpx
+import pytest
 import pytest_asyncio
 from ghga_service_commons.api.testing import AsyncTestClient
 from ghga_service_commons.auth.ghga import AuthConfig
 from ghga_service_commons.utils.jwt_helpers import generate_jwk
 from ghga_service_commons.utils.multinode_storage import (
+    ObjectStorages,
     S3ObjectStorageNodeConfig,
     S3ObjectStoragesConfig,
 )
@@ -33,12 +35,21 @@ from hexkit.providers.akafka import KafkaEventSubscriber
 from hexkit.providers.akafka.testutils import KafkaFixture
 from hexkit.providers.mongodb.testutils import MongoDbFixture
 from hexkit.providers.s3.testutils import S3Fixture
+from hexkit.providers.testing.dao import BaseInMemDao, new_mock_dao_class
+from hexkit.providers.testing.s3 import InMemObjectStorage
 from jwcrypto.jwk import JWK
 
+from tests_ucs.fixtures import ConfigFixture
 from tests_ucs.fixtures.config import get_config
+from tests_ucs.fixtures.in_mem_obj_storage import InMemS3ObjectStorages
+from ucs.adapters.outbound.s3 import S3Client
 from ucs.config import Config
+from ucs.core import models
+from ucs.core.controller import UploadController
+from ucs.core.models import FileUpload, FileUploadBox, S3UploadDetails
 from ucs.inject import prepare_core, prepare_event_subscriber, prepare_rest_app
 from ucs.ports.inbound.controller import UploadControllerPort
+from ucs.ports.outbound.storage import S3ClientPort
 
 
 @dataclass
@@ -107,3 +118,61 @@ async def joint_fixture(
             wps_jwk=wps_jwk,
             uos_jwk=uos_jwk,
         )
+
+
+@dataclass
+class JointRig:
+    """Test fixture containing all components needed for controller testing."""
+
+    config: Config
+    file_upload_box_dao: BaseInMemDao[models.FileUploadBox]
+    file_upload_dao: BaseInMemDao[models.FileUpload]
+    s3_upload_details_dao: BaseInMemDao[models.S3UploadDetails]
+    object_storages: ObjectStorages
+    controller: UploadController
+    s3_client: S3ClientPort
+
+
+InMemFileUploadBoxDao = new_mock_dao_class(dto_model=FileUploadBox, id_field="id")
+InMemFileUploadDao = new_mock_dao_class(dto_model=FileUpload, id_field="id")
+InMemS3UploadDetailsDao = new_mock_dao_class(
+    dto_model=S3UploadDetails, id_field="file_id"
+)
+
+
+@pytest.fixture()
+def patch_s3_calls(monkeypatch):
+    """Mocks the object storage provider with an InMemObjectStorage object"""
+    pass
+    monkeypatch.setattr(
+        f"{InMemS3ObjectStorages.__module__}.S3ObjectStorage", InMemObjectStorage
+    )
+
+
+@pytest.fixture()
+def rig(config: ConfigFixture, patch_s3_calls) -> JointRig:
+    """Return a joint fixture with in-memory dependency mocks"""
+    _config = config.config
+    file_upload_box_dao = InMemFileUploadBoxDao()
+    file_upload_dao = InMemFileUploadDao()
+    s3_upload_details_dao = InMemS3UploadDetailsDao()
+    object_storages = InMemS3ObjectStorages(config=_config)
+    s3_client = S3Client(config=_config, object_storages=object_storages)
+
+    controller = UploadController(
+        config=(_config),
+        file_upload_box_dao=(file_upload_box_dao),
+        file_upload_dao=(file_upload_dao),
+        s3_upload_details_dao=(s3_upload_details_dao),
+        s3_client=s3_client,
+    )
+
+    return JointRig(
+        config=_config,
+        file_upload_box_dao=file_upload_box_dao,
+        file_upload_dao=file_upload_dao,
+        s3_upload_details_dao=s3_upload_details_dao,
+        object_storages=object_storages,
+        controller=controller,
+        s3_client=s3_client,
+    )
